refactor commands more
