+++ print3rs-commands/src/lib.rs
@@ -1,1 +1,4 @@
+pub mod commander;
 pub mod commands;
+pub mod response;
+pub mod tasks;
